Allow selecting a specific port from the main view
From now on, the entry point to WebKitWatcher is a static listview
with all the different upstream ports of WebKit in it, so you can
select one port only and see all the bots associated to it.

Of course, there's also an option to see 'All the bots' at once.